* commit '090e4dc4f6854ff0d2fcc2d420f0b01eca510482':
  Update structs and defines for 64-bit.

+++ rs_hal.h
@@ -51,7 +51,7 @@ typedef struct {
     const void *in;
     void *out;
     const void *usr;
-    size_t usrLen;
+    uint32_t usrLen;
     uint32_t x;
     uint32_t y;
     uint32_t z;
